[alpha] Add player controlled game object.
This commit mainly includes the first basic version of the player
controlled game object's movement implementation. It is to be fine-tuned
via testing at a later stage by altering the Player class's static
variables. As of now, a temporary player sprite has been added for
testing purposes.

In terms of documentation the player class has JSDoc style comments
added to each of its existing methods. Furthermore, the GameObject class
is marked abstract and contains a few interface methods. Attempting to
create an instance of GameObject or calling an unimplemented base class method
in the derived class leads to a new exception being thrown.