[alpha] Destroy asteroids on collision with projectiles.
When asteroids are destroyed by the player's projectiles, two next level
asteroids are spawned unless the next level exceeds the maximum one. The
newly spawned asteroids separate in opposite directions.

Another minor, but important, change is changing the datatype which is
used to store game objects in PhysicsGroup instances from 'Set' to
'Phaser.Structs.Set'. The reasoning behind this change is that the
latter implementation has a rather useful method 'get' which is utilized
in this commit.