[alpha] Add flying saucers.
Saucers are enemy spaceships. They currently only fly and spin in random
directions. Saucers do not interact with any other game objects for now.
Similarly to asteroids, saucers also come in levels (2 as of now). Each
next level of saucer is smaller and faster than the previous.

Note: the uploaded saucer sprites ('.png' files) are temporary. However,
the latter come with the final dimensions of both levels of saucers.