[alpha][bugfix] Fix error on asteroid wave spawn when player is dead.
Previously, when a new asteroid wave was to be spawned, the main scene
attempted to use the player object's sprite's coordinates to create a
keepout zone around the player. However, at that point the player's
sprite is non-existent. Now the player's coordinates are only used if they are
not dead. Otherwise, the main scene defaults to using the world's center.