[full-release] Add bitmap text buttons.
These buttons use a bitmap font to visualize their contents.

The required bitmap font files are stored as assets. They are treated as
large files and are uploaded via git LFS. As of now, there are two new fonts -
'Hyperspace' and 'Hyperspace-Bold', which will be used for the game's UI.