[full-release] Add main menu scene - opening title screen.
When the game is booted up the player is treated to a title screen which
very much resembles the classical Asteroids game's title screen. The
main menu is minimalistic and has only one button which launches the
game.

Changes:
- scene switching: Whenever the scene manager switches a scene (via the
'Game#switchScene' method) not only is the new scene loaded - firstly
the current scene is stopped. This results in scenes being automatically
cleanup up, not leaving any leftover objects;
- hyperspace fonts: Load both the 'Hyperspace' and 'Hyperspace-Bold'
fonts and cache them in the Boot scene;
- main menu: Include main menu scene.